move cookieJar addr to pkg
