+++ modules/pipeline/pipengine/actionexecutor/plugins/apitest/logic/main.go
@@ -22,9 +22,9 @@ import (
 	"golang.org/x/net/publicsuffix"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/modules/pipeline/pipengine/actionexecutor/plugins/apitest/logic/cookiejar"
 	"github.com/erda-project/erda/modules/pipeline/spec"
 	"github.com/erda-project/erda/pkg/apitestsv2"
+	"github.com/erda-project/erda/pkg/apitestsv2/cookiejar"
 	"github.com/erda-project/erda/pkg/envconf"
 )
